pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trans-morph"
version = "1.0.0"
description = "Instant CLI format converter: CSV, JSON, XML, YAML, TOML with validation"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "trans-morph contributors"}]
keywords = ["converter", "csv", "json", "xml", "yaml", "toml", "format", "transform"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "tomli>=2.0;python_version<'3.11'",
    "tomli-w>=1.0",
]

[project.scripts]
trans-morph = "trans_morph.cli:cli"

[tool.setuptools.packages.find]
include = ["trans_morph*"]
